add more functions
添加了更多细节上的功能，比如删除就诊记录，撤销对健康档案的修改
然后完善了健康档案和就诊统计组件中更新值的逻辑，将更新的行为封装成函数，在生命周期函数和其他需要的时候调用
