+++ src/main.js
@@ -5,9 +5,9 @@ import store from './store'
 import ElementUI from 'element-ui'
 import 'element-ui/lib/theme-chalk/index.css';
 import './assets/iconfont.css'
-import myCharts from './myCharts.js'
+// import myCharts from './myCharts.js'
 
-Vue.use(myCharts)
+// Vue.use(myCharts)
 
 Vue.config.productionTip = false
 Vue.use(ElementUI);
